Finished updating the infinite scroll for entities
Summary:
Finished updating the infinite scroll for entities

Updated the object that handles the scrolling.

Merge remote-tracking branch 'origin/DEV-19' into DEV-19

Finished updating the infinite scroll for entities

Finished updating the infinite scroll for entities

Updated the object that handles the scrolling.

Merge remote-tracking branch 'origin/DEV-19' into DEV-19

Removed the routePath solution on getting the list container instead used the React Refs.

Added Scroll Loading indicator to display at the bottom of the table

Test Plan:
Scroll to the bottom of the page and it will load additional 50 more entities.

Scroll to the bottom of the page and it will load additional 50 more entities.

Reviewers: sky

Differential Revision: http://src.aereusdev.com/D15