Fixed unit tests
Summary:
The main reason that the unit tests were being so strange is that entity/definitionLoader has an option for synchronous
and asynchronous opertation. But when you added the code to reuse old request, it got confused and never returned
the synchronous request.

Test Plan: Run npm test and all tests should pass.

Reviewers: marl.tumulak

Differential Revision: http://src.aereusdev.com/D51