Added ts_updated field because some of the entity has ts_entered and some are empty string. Incase there are no ts_entered value the ts_updated will be displayed as a template.
Summary: When checking the new status update dashboard widget, it seems it is displaying incorrect data.

Test Plan:
Steps To Reproduce: 1. Login to Netric Client 2. Add Status Update Dashboard Widget 3. Check on the Status Updates List.
before when one of the  status update item is clicked some of the item will display no date and some are has date. so I added ts_updated in the field in case if the status update entity has a value with ts_entered from the old record to avoid confusion if the user update the status update entity.

Reviewers: sky, marl.tumulak

Subscribers: joemariel

Differential Revision: https://src.aereus.com/D630