Fix the backend issue when the new custom advance search is saved.
Summary: This story is related to Setting default browser view is not working.

Test Plan: Steps to reproduce. 1. Go to work module 2. Click on the search icon located in the top right 3. Click on the ellipsis (this will display the advanced search) 4. Modify the current browser view 5. Hit save button . (on the first time the ID value is string like: my_task) so to avoid the error in saving string type in the ID field which is integer type it needs validation.

Reviewers: sky, marl.tumulak

Subscribers: joemariel

Differential Revision: https://src.aereus.com/D634