Adding some indexes to speed things up in production
Summary: Many of the updates in chat have resulted in a lot more queries that were heavy on the server. These indexes have increased the performance 400%

Test Plan: Test chat module and check to see if it is faster. You can also look at the host status of dat1-prd-lsndfw.aereus.com in grafana.aereus.com. The load used to be 10-50. Now it is more like .5 most of the time.

Reviewers: marl

Differential Revision: https://src.aereus.com/D1493